Refactor the rule to use Patrol Dispatch integration
