--- lib/madePublic/cfnEventSources.js
@@ -1,8 +1,0 @@
-module.exports.eventSources = {
-  eventSources: {
-    webhook: {
-      method: 'POST',
-      apiKey: false
-    }
-  }
-};
