+++ lib/madePublic/dispatchNotify.js
@@ -1,7 +1,6 @@
 const AWS = require('aws-sdk');
-const github = require('../lib/madePublic/event.js');
 
-function notify(event, callback) {
+function dispatchNotify(event, callback) {
   const sns = new AWS.SNS();
 
   const repoName = event.repository.name;
@@ -55,6 +54,4 @@ Consider notify the security team in case of any information leak.
   });
 }
 
-module.exports.fn = (event, context, callback) => {
-  return github.madePublic(event, notify, callback);
-};
+module.exports.dispatchNotify = dispatchNotify;
