+++ package.json
@@ -16,7 +16,8 @@
   "dependencies": {
     "github": "0.2.4",
     "d3-queue": "^3.0.0",
-    "@mapbox/lambda-cfn": "^2.0.0"
+    "aws-sdk": "^2.123.0",
+    "@mapbox/lambda-cfn": "file:/Users/matiskay/mapbox/lambda-cfn"
   },
   "scripts": {
     "test": "eslint **/*.js && NODE_ENV=test tape test/*.test.js test/**/*.test.js",
